src should not be required
How I use this component:

If `this.props.maybeNull` isn't `null` but it fails to get the image, it falls back to `brokenImage`

If `this.props.maybeNull` is `null`, it also falls back to `brokenImage`

```js
src={this.props.maybeNull}
fallbackImage={brokenImage}
```

Alternative I could do

```js
src={this.props.maybeNull || brokenImage}
fallbackImage={brokenImage}
```

But this is a lot more verbose and `fallbackImage` (by name) should take care of both use-cases IMO.

Do you think this is a valid use-case? If not, what do you think should be the right pattern?